FIX BACKEND INABILITY TO CONNECT TO DB:
This is a combination of 2 commits.
  This is the 1st commit message:

WIP STUPID ROOTLESS DOCKER:

Volume inside the container cannot change owner
stupid backend container pretends not to be able to connect to the db if
asked to do so when launched and perfectly does afterwards. I will burn you
tomorrow, I am starving.

 This is the commit message #2:

FIX BACKEND FUNKY BEHAVIOUR:

Migration files are needed to be deployed so they are removed from the
.gitignore and will be pushed to the repo.

DATABASE_URL as it was written in the .env was incorrect and lead to
connection issues similar to what happened earlier with the frontend
container : localhost within the container solely represents the
loopback of the container and does not support outside connections.
Therefore the address of the DB needs to be the container address as it
is defined by docker compose network.

Dockerfile CMD is set to migrate db and start server with the custom
script "npm run start:migrate:dev"