UPDATE PLUGINS & STORE (FRONTEND):
Added a new 'websocket' plugin with its store.
Renamed plugins and stores with the suffix .plugin.ts / .store.ts

The 'websocket' store allows us to monitor the state of our websockets (whether they are connected or not).
And the plugin is there to easily provide the store throughout the whole application.