UPDATE WEBSOCKETS (FRONTEND):
Finally, succeeded in getting the socket through props!!!!
Then, we can instantiate the chat websocket in the default layout instead of the chat page, enabling the connection to be kept active regardless of the app page.
It also avoids to manually handle some tricky disconnection cases.

i.e. this implementation will allow us to add a notification system later on if needed.