REWORK CHANNEL (BACKEND):
Simplify the logic of the join method (channel.service.ts).
Even if the user already has a connection with the given channel, we still need to join the socket.io room for a given socket (not implemented yet).
Then in any case we need to return the newly created or existing ChannelConnection