FIX UNIQUE IDS CHANNELCO && RELATION (CHAT):
!!! THIS BROKE PAGINATION !!!

Reworked the database AGAIN.
Actually we do not need an id for the channelConnection or Relation
models.
What we do want instead is to be able to find one relation or one
channelConnection based on the pair of identifiers (user/channel) or
(user/user).
Well this exists in prisma. @@id was the way to do it instead of
@@unique, and we can also specify a "name: ".

Apparently it's a known issue that types created by prisma do not
include these @@unique or @@id...
Therefore we still need to use deleteMany... that's odd.