REWORK CHATBOX COMPONENT (FRONT):
update chatbox to be at the bottom:

Message input area is at the bottom of the chat page and the messages
are stacked above it.
It was a long battle against ghost containers and flexbox

update chatbox (front):

Chatbox is now a <v-textarea/> which will expend up to 5 rows and be
scrollable (is that a word?) above this limit

update chatbar belongs to chat_page (front):

ChatBar is a component of the chat page instead of being accessible
from every page

update chatbar (front):

Add a vertical divider between the tabs and the windows to improve
readability

update chatbox events (front):

Icon is clickable and sends message, Enter sends message, Enter+Shift
adds a '\n' to the input but does not send message

update chatbox split btw input/display:

ChatBox is a container which encompasses messageInput and ConvDisplay.

MessageInput sends the input to the back, ConvDisplay receives it and
displays it