UPDATE AUTH SERVICE:
Add a generateTotp on authService, it generate a secret, set it to user, generate a qrcode and return an object containing the secret, the qr_code, and the otpauth link.
Add a field on jwt token to store if the user is 2fa logged, this field is set only if user enabled 2fa.
Add a check for 2fa before injecting user on the socket context.