ADD CHANNEL PATCH METHOD (CHAT):
Adding channel update method changing password, kind and name.
Changed the channel and createChannel Dtos to check for password if
kind is set to protected directly in the Dto.

Renaming variables to make it more explicit (channelId, targetId and so
forth)

Added a DirectChannelDto for private messages (next route to be built).