ADD COMMUNICATION WITHIN CHANNEL (FRONT):
Channel store fetches channel messages from back and stores it in a map.
This map in then used to display the messages within the corresponding
convDisplay.
Waiting for aweaver's PR to fix joinroom issue to be able to instantly
display messages between clients

fix joinroom needing socket (front):

Joinroom needs the socket id to send the message to the correct room