REWORK CLIENT MAP (CHAT SERVICE):
Remove the client map (Map<userId, socketId>) from chatService and use built-in socket.io functionality instead.
On a socket connection, we simply join the socket to the `user:${socket.user.id}` room.
In this way, we can easily retrieve each socket for a given user or easily broadcast a message to a specific user using only the socket.io room features.