FIX ISINCHANNELGUARD FOR MESSAGES (CHAT):
I put back the isInChannelGuard that was wrongly removed on the Get
message route.

The idea behind removing it was that it was no longer needed for sending
a message, because the new guard checking if the user is not muted is
enough to check if the user is allowed in the channel.

I shouldn't have removed it from the get message route!