UPDATE USER CONTROLLER AND CHAT GATEWAY (BACKEND):
Update the user controller and chat gateway, in order to implement the basic status logic.
The get(/user/:id) will not return the user and his current status.
The gateway will notify each connected clients with an event `user:${id}:status` with the current status of a player connecting or disconnecting.