UPDATE REQUEST GET INSTEAD OF POST:
I was unsure if we would prefer to use a get or a post method.

A get method was indeed more accurate to comply with the norm.
Though a post would have made sense to use a json instead of a get
route.

We went for the get route on user/name/....
If no user is found throwing an ImATeapotException with message "No such
user" and error "I'm a teapot"