ADD EVENTS FOR PRIVMSG:
Now joining rooms in privmsgs. Also sending event privmsg:create to
update the list of rooms.

The get route on privmsg makes it easier for the front.
Now returning a name to be displayed.

It returns an object { name: string, channel: Channel }

The name is the name of the other person in the channel.
So a priv conversation between Foo and Bar will return Foo for Bar and
Bar for Foo. Got it?!
Also I don't really know where foo and bar came from, I think it's
dreadful.

The Channel contains a list of channel connections and users.